Link all details

See merge request Qmar/pharma!27

+++ lib/core/utils/formatter.dart
@@ -74,8 +74,7 @@ mixin Formatter {
     }
     return false;
   }
- static String formatPrice(double price) {
-    final formatter = NumberFormat('#,###.##', 'en_US');
-    return formatter.format(price);
-  }
+  static String formatPrice(int price) {   final formatter = NumberFormat('#,###.##', 'en_US');
+   return formatter.format(price); 
+   }
 }
